subsys: bootloader: fw_metadata: Add in/out ABIs to the firmware info
Abstract it to an "ABI getter" which is itself fixed-length, and can
retrieve variable length ABIs.

The ABI getter contains a single function pointer which is used to
retrieve the actual ABI in the form of a pointer to a const struct.

The struct contains whatever the ABI consists of, and is assumed to be
known by the caller.

ABIs are versioned.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>