fw_info, bootloader: Explicitly invalidate slot if it fails validation
Add a "valid" member in the fw_info struct which must be 0xFFFFFFFF for
valid images. This value is set to 0 if validation fails for any reason.
This will speed up subsequent checks and also make it obvious that the
image has been invalidated.
Also, add tests for some fw_info funcitonality (invalidation,
fw_info_find, and memeq) and update info struct version because of
breaking changes

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>